refactored code by moving classes into respective packages
