+++ src/main/java/command/AddDeadlineCommand.java
@@ -1,9 +1,4 @@
-package dolla.command;
-
-import dolla.Ui;
-import dolla.task.Deadline;
-import dolla.task.Task;
-import dolla.task.TaskList;
+package command;
 
 /**
  * duke.command.AddDeadlineCommand is a command used to add a deadline task into the duke.task.TaskList.
