+++ src/main/java/command/AddEntryCommand.java
@@ -1,6 +1,6 @@
-package dolla.command;
+package command;
 
-import dolla.Ui;
+import ui.Ui;
 import dolla.task.Entry;
 
 import java.time.LocalDateTime;
