+++ src/main/java/command/AddFixDurationCommand.java
@@ -1,6 +1,6 @@
-package dolla.command;
+package command;
 
-import dolla.Ui;
+import ui.Ui;
 import dolla.task.FixDuration;
 import dolla.task.Task;
 import dolla.task.TaskList;
