+++ src/main/java/command/ErrorCommand.java
@@ -1,7 +1,4 @@
-package dolla.command;
-
-import dolla.Ui;
-import dolla.task.TaskList;
+package command;
 
 /**
  * duke.command.ErrorCommand is a type of command that is generated if a proper command
