+++ src/main/java/command/ViewScheduleCommand.java
@@ -1,6 +1,6 @@
-package dolla.command;
+package command;
 
-import dolla.Ui;
+import ui.Ui;
 import dolla.task.Task;
 import dolla.task.TaskList;
 import java.time.LocalDate;
