+++ src/main/java/dolla/Main.java
@@ -1,6 +1,6 @@
 package dolla;
 
-import dolla.command.Command;
+import command.Command;
 import dolla.task.Task;
 import dolla.task.TaskList;
 import parser.MainParser;
@@ -9,27 +9,27 @@ import java.util.ArrayList;
 import java.util.Scanner;
 
 /**
- * <h1>duke.Dolla</h1>
- * duke.Dolla is a chat-bot styled todo_list manager.
+ * <h1>duke.Main</h1>
+ * duke.Main is a chat-bot styled todo_list manager.
  *
  * @author  Aik Peng
  * @version 1.0
  * @since   2019-07-26
  */
-public class Dolla {
+public class Main {
 
     private TaskList tasks = new TaskList(new ArrayList<Task>());
     private String mode = "dolla";
 
     /**
-     * Creates an instance of Dolla using a data loaded from /data/dolla.txt
+     * Creates an instance of Main using a data loaded from /data/dolla.txt
      */
-    public Dolla() {
+    public Main() {
         //tasks = new TaskList(Storage.load());
     }
 
     /**
-     * Runs the main program of duke.Dolla
+     * Runs the main program of duke.Main
      * @throws Exception when exceptional condition happens
      */
     public void run() throws Exception {
@@ -54,6 +54,6 @@ public class Dolla {
     }
 
     public static void main(String[] args) throws Exception { // Exception needs to be handled?
-        new Dolla().run();
+        new Main().run();
     }
 }
