+++ src/main/java/dolla/Reminder.java
@@ -1,8 +1,9 @@
 package dolla;
 
-import dolla.command.Command;
+import command.Command;
 import dolla.task.Task;
 import dolla.task.TaskList;
+import ui.Ui;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
