+++ src/test/java/MainTest.java
@@ -2,7 +2,7 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class DollaTest {
+public class MainTest {
     @Test
     public void dummyTest() {
         assertEquals(2, 2);
